Better piano samples
These samples are from https://musical-artifacts.com/artifacts/2849 and 
converted to MP3. They need to be credited somewhere in accordance with 
the CC-BY license. 

Some very problematic noise can still be heard on files like 
MaouDamashiiTown10, that devolves into pure white noise near the end. 
The synth seems ill-equipped for polyphony. The samples are top-notch 
this time so the code is clearly the part at fault, but we need to 
figure out why.