Remove play and interpret buttons in favor of single play/pause button
The button sits next to the number inputs for the indices. It's animated 
purely in CSS. 

The current mode of the performer is altered in two places : 

- When the button is clicked (toggling from play to silent or the other 
way around) ; the scroll bar then emits upwards so the MFP vue can 
update the performer's mode
- When any command is detected ; this is done at the app root, and it 
updates the mode itself. To prevent the passive playback loop from 
continuing, it first sets the performer to silent, immediately changing 
to perform, which is not audible. As well, it updates a store variable 
indicating that perform mode has started so that the scroll bar can 
update the play/pause button if need be.

A glitch seems to linger with this new interface where playback will 
sometimes stop during a file, at which point nothing can ever make sound 
except reloading the app ; not changing files, nor changing tabs.