Smoother transition between listen and play state
Instead of rudely interrupting passive playback, wait for the currently 
playing set (start or end) to finish, if it's shorter than a second (if 
not, wait a second). 

This works well, except that sometimes note on/off mangling makes it so 
the set isn't properly played. So I'll have to venture into that area in 
the next commits.

As a QOL, added left and right as shortcuts for moving the scroll bar. 
Only the main scroll bar, though : I don't think many people care to do 
it for the keyboard velocities.