Fix shadow notes by JS monkeypatch
Normally, the C++ lib is set to shift note offs before the same note ons 
in the same set. But the shift mode parameter seems to have been omitted 
and the shift does not happen.

This shifts notes forcefully before playing. For now, this is only done 
for the internal sampler/built-in synth. The shift step will be removed 
once the C++ wasm is updated.

However, this commit is also the occasion to optimize note event 
transmission : they are now passed in batch to the IOCTL in the 
callback, which is observed to put significantly less strain on the 
audio buffer.