Multi-channel built-in synth
Because shifting by pitch only introduced some inexplicable staccato 
(given that the sequence of events returned by the MFP did not actually 
change in the files where this behavior was exhibited), it was decided 
instead to upgrade the built-in synth to properly use MIDI channels.

Note that MIDI channel can play multiple notes (they are polyphonic), so 
each one keeps track of the notes that are being played and stores a 
reference to the audio node playing them. Util methods are provided to 
access that player, set it, and remove it.