Smooth mode transitions around loop boundaries
Sets were still repeated on the first and last indices of a loop, 
because the mode needs to play that when it's on its own, but we need to 
avoid repeating it when shifting. 

Boolean flags are introduced that are set at tricky times to ensure that 
no repetition happens between modes, whether the mode is changed right 
on the boundary, or before it. 

This might still leave a few gaps, but it's been tested to work.

I also took the opportunity to clear out the code for setMode, most 
notably moving playNextSet into a private method.