Fix multi-firing of external inputs
JavaScript's binding is just so exceptionally bad. 

We have to bind the listener to this for it to work, which is bad enough 
; but binding it on the fly copies the function, and so the listener was 
never removed.

So we have to create a constant with the reference of the listener bound 
to this, which is what it should be in the first place.