Save last output and bypass loading screen if not synth
This is done through the local storage to persist through page 
refreshes.

To prevent selecting the built-in synth when it hasn't been fully 
loaded, the corresponding option in the output selection is disabled 
until load is complete.

Also, factorized the default IO ID and amount of keys throughout the 
app.