Fix speed display on file upload
Because the scroll bar has no "speed" prop (and it would not make sense 
to create one as of now, since the component is also used for sliders), 
the value of the speed input is not reset when a file is uploaded, 
despite the MFP itself resetting it behind the scenes.

This break in reactivity is only superficially fixed here. The real 
solution is to extract the slider component entirely and divorce the two 
parts' functions. This way, the scroll bar can have a speed prop without 
muddying the API.