Issue 31 (bis) : Velocity profile for variations across a piece
Chord velocity mappings are local : for the same input velocity, they 
produce the same output. This means that previously, perform velocity on 
passive playback eliminated all velocity variations in the file. 

This is fixed with the use of the chronology's inner container, 
profiling the max velocities across the piece, to then readjust the 
velocity to use with each new command.