Change InterruptionGuard logic to allow for grace period
Previously, it was common to come in too late after a note was triggered 
when trying to trigger it with perform mode. Now, before half of the 
note's duration has passed, the InterruptionGuard will set flags that 
will make the performer act as though the input that triggered perform 
mode also triggered the note in question. 

The duration of this grace period might be adjusted in the future ; 
perhaps to a quarter of the duration ? A third ? Perhaps it should be 
shorter when the duration itself is shorter ? (i.e. : if dt is only, 
say, 50ms, maybe it shouldn't matter if you come in after 25 or 40). 
Either way, for now, this is a net improvement.