Fix issue #49 : only restore I/O if it's still available
...but this required a few other things too.

The app root was not awaiting on the input setting operation from the 
IOCTL, so the undefined check could not be reliably performed on the 
provided input or output ID. 

updateInputsAndOutputs is now async and awaited on by the root's 
created() method.

This would suffice, but there is a Firefox-specific issue : when 
rejecting MIDI access, if Firefox finds no devices, it will wait a 
random length of time. Thus, we must wait along with it, or the app root 
will never be updated on the sample loading process, and will remain 
stuck. 

The loading screen now features a loading message for the MIDI access 
request. It is unreasonably long (on top of being random) for Firefox, 
and instantaneous for all other tested browsers.