"Fix" device refresh when unplugging
When refreshing devices, if the current selected input or output is no 
longer available, restore the default. 

This requires a bit of handiwork to ensure event handlers remain 
unduplicated : store a set of unplugged devices whose handlers couldn't 
be removed, and remove them all at the next refresh. 

This leaves a bug where the computer keyboard still functions when 
plugging in an external device again, but since our endgoal is to have 
multiple device input, it's not much of a problem ; besides, we'll come 
back to this issue later.