First (buggy) attempt at piano roll using HTML-Midi-Player/Magenta.js
The PianoRoll component is used as a wrapper for the HTML-Midi-Player 
Visualizer element, which is itself a wrapper of Magenta's 
PianoRollCanvasVisualizer. 

Magenta works using sequences of note objects that have start times and 
durations, instead of note on and note off events. This paradigm change 
forces us to store this alternative representation at parse time. 

After this is done, a map is constructed that lists each note/channel 
pair's occurrence with its start and end times. This is used to finally 
construct an array of pseudo-Magenta note objects that can be used to 
anchor visualizer synchronization on each set boundary.

Still, this isn't enough. Magenta's MIDI parser leads it to have 
different times from ours ; moreover, to automatically scroll through 
the visualizer when it's activating a note out of sight, Magenta uses 
===, requiring us to use their own objects, not ours.

As a result, the pseudo-Magenta object used to trigger the visualizer 
redraw is instead compared by start time to all possible occurrences of 
the pitch, for one that starts less than 20ms apart. The corresponding 
Magenta note is then fed to the redraw method. 

This is a very arbitrary threshold that does NOT work for many files, as 
they break down in the middle or refuse to work entirely.

As a consequence of this very empirical approach, several more problems 
remain : 

- In general, there are a lot of times when the threshold will fail to 
detect the right note occurrence
- The whole play-time operation (of comparing pseudo-Magenta and Magenta 
notes to find the right one) slows down playback, not significantly, but 
noticeably
- Sometimes, the visualizer will not scroll, presumably because a note 
is being held that prevents it from detecting a boundary change
- Sometimes, non-overlapping notes are all highlighted at once (see : 
MLS, index 354)
- ...and likewise, sometimes the number of highlighted notes does not 
even match the number of notes in play
- ...notes can even be highlighted multiple times by mistake
- The visualizer does not scroll back to the beginning, on loop nor file 
upload (calling the provided reload() method does not change this)

These bugs are not very detrimental to our situation, because this is a 
proof of concept : the end goal is to adapt the Magenta visualizer to 
our needs, integrate it directly, and obtain a single source of absolute 
time information through the lib itself.

Miscellaneous :

- Updated webpack.config.js because Magenta requires polyfills not 
automatically provided by webpack, and to allow for HTML-Midi-Player's 
midi-visualizer native custom element