Fix visualizer (almost), but it's slow and JS is bad
Using startTimes as a comparison anchor is just an unsustainable idea. 
Now, we rely on the order of appearance of notes in a file, which is 
dictated by the start dates.

At parse time, no absolute start and end times are stored anymore. 
Instead, we wait until the performer has been fed, and then iterate over 
the starting sets. For each note on, we increment an index which starts 
at 0 and is unique for every pitch (regardless of channel). Then, we 
send to the callback the combination of this pitch and the index for 
every *last* note in every set.

The callback can then look through the note sequence (specifically : the 
notes with the same pitch in the note sequence) and pick that which is 
found at the given index. This way, no duration has to be approximated, 
and all discrepancies are erased.

But this still doesn't work.

For one : some files still show odd behavior. Although they will now all 
scroll forward and at least one note is guaranteed to highlight on every 
set, Magenta's parsing appears to have approximation issues between its 
own notes, because not every note in the set will always light up. 

As well, we can sometimes observe highlighting jump back one set (never 
more). This doesn't hinder progress too terribly much, as everything 
still moves forward. It's still strange, and something to look into.

The scroll still does not move back to the start on loop nor on file 
upload. This is unfortunately a limitation of Magenta itself : the 
official demo for HTML-Midi-Player suffers the exact same problem (it 
reloads the entire page to work around this on upload).

But by far the worst problem is that Magenta's parser *does not* in fact 
sort notes by startTime. They are most likely left in the exact order in 
which they are inserted track by track. This means that we have to 
perform the sort ourselves, and unfortunately for us, sorting an array 
in JavaScript appears to copy the objects in that array. 

This means that the sorted array can no longer be used to obtain the 
original reference that the visualizer uses to check for scrolling. So, 
for scrolling to work, a *second* operation must be performed to look 
for the identical object in the base collection. On every redraw 
callback. The larger the file, the worse this gets.

There's still a long way to go.