Use sort in place to filter and avoid double find, still suboptimal
Reuse the filter method instead of a preconstructed map, but on the 
sorted list. This avoids the tedious double find and improves the 
performance situation mentioned in the previous commit, but using the 
map (as done in the commented code) would be better. It should work. I 
don't understand why it does not.