Reproduce Magenta PianoRollSVGVisualizer into own SFC
To avoid packaging HTML-Midi-Player and, through it, Magenta, which are 
heavy dependencies, the code of the Magenta visualizer has been 
reproduced in the PianoRoll.vue SFC, with some small adjustments, and 
credit given to Google. 

Because notes with a start and end time work better to determine 
overlap, for now, the SFC converts the outgoing chronology into an array 
of timed notes (vestigially called "note sequence"). 

On redraw, the SFC receives the index of the current set, which it 
matches against the index of the first note in the set.