Limit hover highlight to silent mode and disregard overlaps
The user should not be able to highlight notes while performance or 
playback are doing so. Therefore, it is now only active in silent mode.

This was harder than expected, because our current tracking of mode is 
very flimsy, and should be centralized in the store as soon as this 
branch is merged. A special event chain was the only way I could find to 
get this to update properly for now.

As well, the hover highlight no longer includes notes that are 
overlapping with the highlighted set, because this visual cue is meant 
to show the user what notes will play once they move there, and notes 
that started before a set are not played when we move there.