Prevent highlight of non-played notes even in perform/playback
This measure was already taken care of for hover highlights, but perform 
and playbacks were also highlighting notes that aren't playing when the 
index was jumped. Now, the piano roll is informed every time the index 
is set by the user, and it defines that index as a threshold : no note 
that starts before that point is highlighted. This works even on loop 
(I'm not exactly sure why, but it does).