Briefly play clicked set and scroll visualizer with index changes
Using getCurrentSetPair, the current set's sound can be briefly played 
when the user clicks on a visualizer note. This, of course, only happens 
if the performer is currently silent. 

As well, the piano roll now scrolls into view as soon as it is notified 
that the index has been updated. This way, the scroll bar and the piano 
roll are now in permanent bijective synch.