Experiment with permanent scrolling with active notes in the middle
This foregoes any condition for scrollIntoView and dicates that the 
active note must always be in the middle (it is only on the boundaries 
if at the start or end of the file). *

This prevents jarring screen changes, which are hard to keep track of, 
but introduces two problems : 

1. Smaller notes tend to "blink" in the redraw operation
2. When the time step (the duration of the sets) is not very short, the 
scrolling seems janky and inconsistent (because it's scrolling on a set 
basis). 

The second issue would suggest we need to constantly scroll by the 
smallest time step. The first would, on the contrary, suggest that we 
need to scroll less often to mitigate redraw overhead. 

This is a predicament.