Fix scroll on click (again) (and it's still not done)
Clicking notes still moved the window to have that note in the middle. 
This is no longer the case ; now, any note clicked in the window will 
not trigger a scroll.

Moving the index through the scroll bar will also now move the view to 
use the chosen set as its left boundary, rather than locating it in the 
middle.

This behavior remains inconsistent :

- On start change, the index is not updated and thus the view doesn't 
scroll (this oversight was already present in 7ef88c5, i.e. before 
middle scroll) 
- On loop with a non-trivial ( => beyond container_width / 2) start 
index, the loop places the starting index in the middle of the window 
again, because it's not a user-initiated jump 
- The most important : if the user clicks on a note after the middle 
point of the window, starting playback or perform will make the window 
jump to have it in the middle immediately. We need to increment 
scrolling until it is in the middle and then reprise normally.