Scroll piano roll on scroll bar start and end flag drag
As well, correct my own oversight mirroring Magenta's, that prevented 
proper backwards scrolling.

By necessity, scrolling smoothness is asymmetrical : scrolling forward 
only moves the window to realign the new left border, because we choose 
not to realign with a note already in the window (so as to not realign 
on click).

Conversely, scrolling backwards is smooth, because the index will scroll 
to will always be the first one behind the left border as we go. I don't 
think there is a solution to this.