Dynamic highlight and drastic performance improvement
The PianoRoll now mapStates the store's keyboardState to get the mask of 
each channel. From that information, it can know if a note is actually 
playing when examining its activity.

It also stores the startTime of every occurrence of a note/pitch 
combination.

Finally, the MFP.js emits redraw messages on BOTH start and end sets 
when in perform mode (playback can't step out of line from the file, so 
it's not needed).

This allows the visualizer to only repaint notes if they are actually 
playing. In other words, it syncs up with the user's input. If a note 
should be playing but isn't, it's not highlighted ; if a note isn't 
overlapping or synced, but is the last occurrence and is indeed shown to 
be playing by the mask, it is.

This was supposed to be the whole commit, but it also contains a 
significant boost to the visualizer's performance, seeing as it lagged 
behind significantly ever since the SFC reimplementation. This is 
accomplished through the following changes : 

- When redrawing, only iterate through a small portion of the notes, not 
the entire sequence. The leftmost note is the left border ; the 
rightmost note is the current note. For now, this prevents highlight of 
user-held notes on short (non scroll-inducing) loops when the loop 
rewinds. This should be fixed soon.
- To alleviate the calculation of the boundaries, store the sets' end 
indexes (besides the starts, which were already stored) and every 
rectangle's right edge's X position on parse (in reverse order, because 
the lookup is for the last matching index, and findLastIndex is 
significantly slower than findIndex).
- Make containerWidth, an information required on every iterated note of 
a redraw, a data property stored on DOM update, in order to eliminate 
the overhead of querying it every time 


Miscellaneous : 
- Fix bug in onNoteClick() where the last set would not react properly 
due to findIndex returning -1
- Change the name of "playOrPerform" in isPaintingActiveNote to use that 
name as a convenience alias in redraw instead
- Add a null check in onIndexJump that allows for proper re-mounting