Separate redraws, improve performance, bugfix, retain dynamic highlight
As previously described, the window-restrained iterative solution for 
highlight is insufficient for some edge cases. Therefore, it could not 
remain in place.

The for loop approach through the entirety of notes is fundamentally 
wasteful. I don't know and don't care much to know how and why Magenta 
gets away with it with very little computational loss, but with the need 
to fix the aforementioned edge cases, it was finally decided to drop it 
altogether.

The two different cases of highlight (for perform/play, and for 
hover/click) are now separated into dedicated methods : refresh() and 
paintNoteSet(). Both use the same painting submethod, fillActiveRects(), 
which takes, as inputs, the notes to fill, and both calculate these 
notes for them. Neither of them use isPaintingActiveNote() anymore, 
leading its deletion.

refresh() uses a map replacing and expanding upon lastOccurrences(). 
Comparing the keyboard state to that map and the current set, it 
determines which notes are currently being played, with no more 
distinction between syncing, overlapping and artificial prolongation. It 
updates the map and sends its values to fillActiveRects(), before 
calling a wrapper of scrollIntoView() called scrolltoSet(), which is 
also used in onIndexJump().

paintNoteSet() simply takes the set of the given note as the notes to 
paint, and passes it to fillActiveRects(). No scrolling is involved.

With this, the visualizer is the most performant it has ever been. Tests 
on other machines should serve to confirm that this is the way to go. 

Misc : 

- Notes in the noteSequence now carry their own index for convenience.
- Changexd click highlight duration from 100 to 200ms.