Materialize sequence bounds in piano roll with vertical bars
Draggable vertical bars (blue, not red, see issue #58) that sync both 
ways with the sequenceStart and sequenceEnd.

Some adjustments to the width and spacing had to be made, and things are 
still a little rough. At times the drag behavior becomes unresponsive or 
completely broken and I have to reload the page, but this may not be 
reproduced, so for now, I am committing it, and we will watch for future 
occurrences.

There are also some kinks to be noted with how the pseudo-drag is 
implemented (SVG doesn't have actual drag, of course) : 

- mouseup and mouseleave have been attached to the SVG. We don't have 
direct access to the surrounding page, and it would be unwise to attach 
them there, so on corners of the svg, the position can be finicky.
- The boundary bars do not cause scrolling (for now). 
- An ugly +1 added to the width of every note.