Correct chronology conversion to account for multiple note ons
In files where note ons can stack without an off, the piano roll would 
produce incorrectly short notes. This is no longer the case, and aside 
from possible remaining hanging notes due to the underlying library, all 
notes should now be drawn just as long as they stay on in the file. 

During the chronology parse, each note on is counted. If the note map 
already has the pitch/channel combination, the following steps are taken 
: 

- If the new occurrence of the combination is another note on, push the 
previous on to the note sequence and increment the counter 
- If it is a note off, push the note stored in the map ONLY when the 
counter has reached one. For each note off, decrement the counter by 
one, and when it has reached one, delete the note from the map (as was 
previously done on every off). 

What this amounts to is that for X note ons in a row without an off, 
only the last one is terminated by offs, and it will take exactly X offs 
to actually terminate it. This is how such files seem to work : MLS is 
the prime example. If there are other cases, I have not yet seen them.