Linked click note release to mouseup
A bit dirty, but it works : the mouseup listener calls the mouseleave 
listener with a null event, preventing the nulling of the registered 
rect, but unpainting the rectangles and emitting a stop signal.

The MFP.js is no longer used for any of this. Note that this means that 
anything played will now conform to the visualizer, not the renderer, 
hence if any visualizer bugs happen in the future, they will be 
reflected in click-play.