Layout changes and resize through wheel
Moved the piano roll on top of the MFP page, displacing the IOManager 
and file input below the scroll bar. 

Set the max height of the piano roll to 60% of the app-content, and 
supported for the piano roll to grow through wheel events with Ctrl held 
down. 

Growing is simply performed via a total redraw, which should not, and 
does not seem to, cause any effects. 

Misc : 

- Atomized the steps of clear() for factorization
- Change getSize() to setSize(), for the same purpose