Hide transitory MFP state being loading screen
Add a "generic condition" prop the loading screen that triggers a 
generic loading message and serves to hide the MFP page as it's 
transitioning between two files. The condition is fed using a reactive 
boolean flag. 

For this to work, the MFP.vue template was encapsulated into a 
loading+MFP dummy container and the MFP is created even when the loading 
screen is active so it can be prepared ahead of time. 

The loading screen was also fitted with dynamic ellipses that serve as a 
pseudo-progress bar. Unfortunately, these only work on the loading 
screen's initial mount, and no successive mount after that : the 
setInterval function used to program the ellipses update callback 
completes as expected in the mounted hook, but the callback itself is 
never called on any later mount. For this reason, only the MIDI access 
loading message has dynamic ellipses, and the generic message shown for 
file loading does not.