setTimeout hack to ensure scrolling to beginning on file load
After the last loading screen commit, the piano roll container now 
receives a scroll event of unknown origin with isTrusted set to true 
*after* the updateNoteSequence function exits and *after* the indexJump 
method has scrolled it back to the start. This forces me to delay 
another force reset of the scrollLeft value using setTimeout, a hack 
which will inevitably fail to work in some environments, but to which I 
see no alternative at present.