First version of MusicXML support
Uncompressed MusicXML files can now be uploaded to the app just like any 
MIDI file. They will be translated into a JSON representation with the 
musicxml-interfaces lib, and converted from there into a MIDI JSON 
identical in format to the one produced for MIDI files. The process then 
resumes at the MFP.js parse step as it usually would.

The parser accounts only for a very small subset of MusicXML : that 
which is necessary for the purposes of our app. It is still very 
rudimentary and will receive enhancements moving forward. 

Some caveats apply to the file input process in MFP.vue in order to 
transparently unify inputs between the two file types : 

- A separate verification process considering only the suffix for 
MusicXML files, as they lack binary signature, being text files. 
- Storing the extension of the file for truncated file name display
- Checking on mount if the stored "buffer" of the mfpMidiFile is of the 
ArrayBuffer prototype, as we directly store the converted MIDI JSON in 
the case of a MusicXML file 
- Moving the set operation of the loading flag before file reader 
operations, in order to ensure that the entire MusicXML parsing pipeline 
is covered under the loading screen as opposed to only its MIDI JSON 
parsing ; this necessitates a call to Vue's nextTick() to avoid timing 
issues