Support compressed MusicXML (MXL) files
These are the most common format of MusicXML files available, as an 
uncompressed MusicXML file is very large in size due to its text-based 
nature.

This is simply a matter of reading the MXL like a ZIP, parsing a 
container file and reading the texts contents of the root file 
registered in that container.