First version of OSMD support for MusicXML sheet music display
The OSMD object is wrapped by a Vue component that provides its
localized container. To ensure proper render, v-show had to be removed
from the MFP container and replaced with pos absolute + z-index +
opacity to hide it, otherwise the container's width would be 0, and
unlike the piano roll we can't just delay drawing until the update
lifecycle hook.

The store's MFPMidiFile now contains an isMidi boolean, and a
musicXmlString space for storing the text contents of a loaded MusicXML
file (which is left blank in the case of a MIDI). This way, the
SheetMusic.vue watcher can read directly from it and MFP.vue's load
logic never has to interact with SheetMusic directly. isMidi also
replaces isTrueBuffer.

loadMfpMidiFile() has been removed from MFP.vue, to displace its
silencing and state reset operations to another step of the process,
encapsulated in method preparePerformerForLoad(). This also means the
mounted() hook sets the loading flag by itself.

fileContents no longer reads MXLs as text, and always contains the
MusicXML string whether the file is compressed or not.

For now, the piano roll is commented out. The next commit will enable
both it and the sheet music display.