Fixed-size visualizers (piano roll and sheet music)
This basically means : nothing changes for sheet music, but the SVG of the piano roll is always at least as big as the container, and the Y value is calculated differently. 

This also means the PianoRoll doesn't use v-show anymore, so emitting "ready" is no longer necessary and the storing of containerWidth can be done at mount time.