Draggable cursors
It *can* be done !! 

By saving the coordinates of the cursor on each set, we can then seek the closest set position on a mousemove event. This allows us to drag *any* cursor, with the caveat that they are originally set to stay behind the score, such that the z-index of start and end have to be increased, and the main cursor is triggered indirectly through the container, as it has to stay behind to not cover the notes it passes over. 

Misc : 

- Fix an async issue that could trigger the sequenceStart/End watchers before drawn is set to false by updateScore()
- Rename every instance of "position" to "date", to avoid ambiguity with graphical coordinates
- Remove switch statements to determine cursors, replacing them with an array of cursor names