OSMD note click and hover support (BUGGY)
OSMD notes can now be clicked to play their sound and move the index, and highlighted both then and on hover, in a manner identical to the piano roll. As a foreword : this introduces severe duplication between the two components that I see no easy way out of, because Vue has no inheritance framework for components, though perhaps mixins could mitigate the issue.

To get access to the notes' graphical elements, another operation is undertaken at parse time, which matches every graphical note with its equivalent in the note sequence.

For this operation to be possible, the MusicXML parser now transmits, alongside tempo events, a complete list of each track's channel changes. 

The loop then adds listeners to every note's SVG root, with a utility method able to find the notehead from any element of its hierarchy (stem, flag, accidentals, etc.) if it becomes the target. This notehead is then registered in two maps : once as key, once as value, against its equivalent entry in the noteSequence.

This allows the listeners to extract noteSequence data by examining their target's notehead ; vice-versa, the painting methods are able to take note sequence notes as input, and deduce the graphical noteheads from there. 

This will also be the basis of dynamic highlighting during play : the idea is to transmit the active notes and simply get their equivalent GNotes, making them the newly painted notes on the sheet. 

The listener and painting logic, though greatly duplicated, still remain unique in their own ways. Most notably : 

1. Because the noteheads are not direct children of the canvas, we cannot simply rely on a querySelectorAll() to see which notes are painted. This means keeping them stored in dedicated arrays, and offers a basis for transition to issue #74 : notes highlighted by the mouse on click and hover, and notes highlighted by refresh will be stored separately. 
2. Conversely, the unpaint() method, equivalent to unfillActiveRects(), takes a boolean value telling it whether it should unpaint all notes, or just the ones that were highlighted by the mouse.
3. The unpaint method is still called by paintSetOrNote() (which calls no subroutine of its own here), but with its boolean arg set to true only if the paint job is for refresh.
4. onNoteUnclick() calls onNoteLeave() by passing its argument, not with null, because onNoteLeave() always needs an event target.
5. It is necessary for mouseover events dispatched by control key checks to have bubbles set to true, because the registered noteHead has no listener : only the root. 
6. Perhaps most noticeably : clicking on a note temporarily hides the cursor, so the highlight is visible. This is the best compromise currently available : the only other option would be to use a different color, either adding to the unstable palette or not high enough in contrast to stand out from the cursor. 

THIS IS NOT THE END. THIS STILL DOESN'T WORK FOR ALL FILES. In some cases, the supposedly bijective maps are not in fact bijective, which prevents some notes from reacting ; this needs to be investigated. In other cases still, equivalents are never found for some notes, halting the process entirely with a dedicated error : this coincides with points where the cursor skips a set (see issue #67), and is thus most likely out of our control. 

Misc : 

- Factorize eligibility for cursor anchoring into a function called isEligibleMfpNote(), also used in the noteHead setup operation 
- Add stop() function to SheetMusic, which calls unpaint with allNotes = true.
- Add clear() wrapper to clearView() in SheetMusic that clears some inner state as well.
- Move SheetMusic clear() operation from watcher directly into updateScore()
- Moved setupCursors next to cursorAnchors and noteHeads setups
- Rename highlightedNote to rectUnderCursor in PianoRoll for consistency between the two visualizers
- The cursor is now shown from the start. 
- Remove comment about cursor track breaking, but the issue is still very much open ; tab switching breaks a lot of things still.