Fix note click/hover bugs by accounting for single-stem clusters
A single SVG <g> element can (as its name indicates) *group* multiple notes, more specifically multiple noteheads, together.

This means that, when examining an SVG family, we should always return an array of noteheads ; and this comes with several consequences.

Most are rather easy to foresee : the listeners have to change ; most notably the noteHeadsUnderCursor property also has to move to an array, with ctrlKey watchers dispatching events to every one of its elements.

However, one particular consequence is extremely disruptive : the noteheads grouped under one <g> can include tie-prolonging notes, i.e., notes that do not actually exist as note events and part of the note sequence ; these must then be filtered out ; but there is no direct way to know which noteheads these are.

This forces us to change the way we setup noteheads. Instead of filtering out GNotes that are tie-prolonging, we include them in the loop, and once we encounter one, skip its corresponding head. This relies on an index value, because GNotes are encountered in the same order as noteheads are returned : from bottom to top of a <g>.

This incidentally also means that every listener must find the first *valid* notehead in the array to know which note and set index to proceed with. 

With this, the only time setupNoteheads() fails should be as a result of OSMD issue 1540 (corresponding to our issue 67)