Transfer velocity sliders into settings popup, and LOTS of CSS changes
If I could, this would be two commits, not one.

The main task here was to put velocity sliders into the settings. This just means taking out all code related to them from MFP.vue and adding them there. But in practice, this highlighted several CSS issues with the app.

First, of course, the ScrollBar component, which is intrinsically broken in its double use. I moved the lone index input before the bar and the reset to align with it so they all stand on a single line. The sliders themselves have been reduced in width and had their proportions altered.

Following this, the number inputs. They have been made to always extend to 100% of their container, and the ScrollBar sets their exact width, so they are identical between browsers.

As well, the font setting was broken and resulted in different fonts for different browsers. The font used is now Firefox's default, across browsers, unifying the layout.

Of course, the popup CSS has been changed. The popup is now top-level of the MFP component, not inside a container. This means its absolute position is actually set in reference to the whole app, allowing for unified centering. Simultaneously, its background was set to cover the whole app.

Finally, the default centering applied to the MFP.vue page ceased to exist as a result of top-level settings, so it had to be re-set individually on the component and the popup elements.

------------------------------

With this : what was this commit actually about ? The settings data model.

Settings are stored as *separate fields* in the store. A *getter* is defined to pack them into a single object that the settings SFC uses. It will then write *back* a single object to the store, which updates its fields individually.

The settings SFC maintains a copy of the object it was passed thanks to a new JS baseline for deep cloning. It will thus be able to detect any changes relatively easily.

Note that, since the popup is only mounted *once*, its inner state has to be reset *manually* on open. This is rather ugly, but compared to the whole CSS mess, it appears clean in retrospect.