Modify keyboard velocities through popup
The popup does something ! WOOO !

Four buttons have been added : apply, confirm, reset (to previous settings), default. 

A watcher has been set up on keyboard velocities in the IOManager, which updates the IOCTL as the store's value changes. 

For the "default" button to work, default settings have been created as a JSON file. It is imported by the settings popup, of course, but also the IOCTL itself and replaces the defaultVelocities previously exported by the IOCTL and injected elsewhere.

In doing this, the localStorage no longer stores the last used velocities : this is because in the next commit, it will store the stockings *as a whole*.