Add IOManager to options popup
The IOManager no longer directly modifies the IOCTL input and output by calling its methods on form change, but rather waits for apply/OK through a watcher, just like with the keyboard velocities.

For this to work, the MIDI init of the IOCTL is also done by the IOmanager (else we run into an async issue with the app root). 

All references to locale properties "input" and "output" have been erased. Saving the preferred input and output now happens through settings.