Change input model to multiple inputs (issue #34)
This means inputIds is now an array. It is managed using a group of checkboxes instead of a select dropdown. 

The logic is mostly intuitively derived from the single-output model. However, there is one exception : the IOManager needs to detect disconnected inputs ahead of time and transmit a force-write instruction to settings to save that change in the local storage. 

Aside from this, there is also the issue of localization : the IOCTL changeLocale(), which was definitely not viable, has finally been dealt away with ; the IOManager itself now handles these labels.